test(ui): split the log panel tests into their own file
Log rendering, pinning, clearing, and unread tracking are exercised together
and belong beside the source file that now holds them.

Tests move unchanged.
